Began V4, will not complete until engine ver works.
V4 is essentially an offload of Abstract Realm's vulkan backend but fitting inside the restrictions of header only (with exception to the static allows for singletons...)

I'm hoping it could be an easier to understand boostrap than whats available from traditional sample sources.

I eventually want to also add support to allow for imgui to be hooked up to it and operate using this backend (my engine will but most likely not straight up to vulkan like V4 may).